+++ crates/lcode/src/app/info/cmds/keymaps.rs
@@ -4,20 +4,20 @@ use ratatui::widgets::{ListItem, ListState};
 #[derive(Debug)]
 #[derive(Default)]
 pub struct KeymapState<'key> {
-    pub keymaps_state: ListState,
-    pub keymaps_items: Vec<ListItem<'key>>,
+    pub items:      Vec<ListItem<'key>>,
+    pub list_state: ListState,
 }
 
 impl<'key> KeymapState<'key> {
     pub fn new(keymaps_items: Vec<ListItem<'key>>) -> Self {
         Self {
-            keymaps_state: ListState::default(),
-            keymaps_items,
+            list_state: ListState::default(),
+            items:      keymaps_items,
         }
     }
     pub fn trigger(&self) -> bool {
         let a = self
-            .keymaps_state
+            .list_state
             .selected()
             .unwrap_or_default();
         if a == 0 {
@@ -27,28 +27,28 @@ impl<'key> KeymapState<'key> {
     }
 
     pub fn first(&mut self) -> bool {
-        self.keymaps_state.select(Some(0));
+        self.list_state.select(Some(0));
         true
     }
     pub fn last(&mut self) -> bool {
-        self.keymaps_state
-            .select(Some(self.keymaps_items.len() - 1));
+        self.list_state
+            .select(Some(self.items.len() - 1));
         true
     }
     pub fn prev(&mut self) -> bool {
-        let i = match self.keymaps_state.selected() {
-            Some(i) => (self.keymaps_items.len() + i - 1) % self.keymaps_items.len(),
+        let i = match self.list_state.selected() {
+            Some(i) => (self.items.len() + i - 1) % self.items.len(),
             None => 0,
         };
-        self.keymaps_state.select(Some(i));
+        self.list_state.select(Some(i));
         true
     }
     pub fn next(&mut self) -> bool {
-        let i = match self.keymaps_state.selected() {
-            Some(i) => (i + 1) % self.keymaps_items.len(),
+        let i = match self.list_state.selected() {
+            Some(i) => (i + 1) % self.items.len(),
             None => 0,
         };
-        self.keymaps_state.select(Some(i));
+        self.list_state.select(Some(i));
         true
     }
 }