+++ crates/lcode/src/app/select/mod.rs
@@ -49,8 +49,8 @@ impl<'tab0> SelectQS<'tab0> {
             ..Default::default()
         }
     }
-    pub fn update_percent(&mut self, cur_perc: f64) {
-        self.sync_bar.update(cur_perc);
+    pub fn update_percent(&mut self, perc: f64) {
+        self.sync_bar.update(perc);
     }
     /// refresh `filtered_qs`
     pub fn filter_by_input(&mut self) {